requirements.txt
Django==5.2.1
djangorestframework==3.16.0
Markdown==3.8
sqlparse==0.5.3
django-filter==25.1
psycopg2-binary==2.9.10
urllib3==2.4.0
django-extensions==4.1
gunicorn==23.0.0
